Code 6
Problem

If you wanna party, if you, if you wanna party
Then put your hands up

Chef wants to host a party with a total of NN people.
However, the party hall has a capacity of XX people. Find whether Chef can host the party.
Input Format

    The first line of input will contain a single integer TT, denoting the number of test cases.
    Each test case consists of two space-separated integers NN and XX — the total number of people and the capacity of the party hall.

Output Format

For each test case, output on a new line, YES, if Chef can host the party and NO otherwise.

Each character of the output may be printed in either uppercase or lowercase. That is, the strings NO, no, nO, and No will be treated as equivalent.
Constraints

    1≤T≤1001≤T≤100
    1≤N,X≤101≤N,X≤10

Sample 1:
Input
Output

4
2 5
4 3
6 6
10 9

YES
NO
YES
NO

Explanation:

Test case 11: Chef wants to host a party with 22 people. Since the capacity of the hall is 55, he can host the party.

Test case 22: Chef wants to host a party with 44 people. Since the capacity of the hall is 33, he can not host the party.

Test case 33: Chef wants to host a party with 66 people. Since the capacity of the hall is 66, he can host the party.

Test case 44: Chef wants to host a party with 1010 people. Since the capacity of the hall is 99, he can not host the party.